Fixed, renamed to dead_chat.py
Renamed deathchat command to dead-chat (dead_chat internally)
Renamed deathchat.py cog to dead_chat.py
Fixed typos
Fixed command since it would not ping everyone but only the @everyone role of a specific server
